dbPath only required if application.conf does not have TILEDB_URI value
