+++ src/main/kotlin/net/maizegenetics/phgv2/cli/LoadVcf.kt
@@ -6,6 +6,7 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.validate
 import htsjdk.variant.vcf.VCFFileReader
 import net.maizegenetics.phgv2.utils.inputStreamProcessing
+import net.maizegenetics.phgv2.utils.verifyURI
 import org.apache.logging.log4j.LogManager
 import java.io.BufferedInputStream
 import java.io.File
@@ -160,62 +161,6 @@ class LoadVcf : CliktCommand(help="load g.vcf and h.vcf files into tiledb datase
             }
         }
         return vcfSampleList
-    }
-
-    // The uri should either be gvcf_dataset or hvcf_dataset
-    // The user determines the parent folder name where these datasets live
-    // The actual tiledb dataset names are constant and are either gvcf_dataset or hvcf_dataset
-
-    // We are only verifying, not creating the datasets
-    fun verifyURI(dbPath:String,uri:String): Boolean {
-        // Check that the user supplied db folder exists
-        check(File(dbPath).exists()) { "Folder $dbPath does not exist - please send a valid path that indicates the parent folder for your tiledb datasets." }
-
-        // Check if the dataset exists
-        val dataset = "${dbPath}/${uri}"
-        val datasetPath = Paths.get(dataset)
-
-        if (File(dataset).exists() && Files.isRegularFile(datasetPath)) {
-            throw IllegalArgumentException("URI ${dataset}is a file, not a tiledb dataset folder.  The parent folder must not contain any files/folders named gvcf_dataset or hvcf_dataset that is not a tiledb created URI")
-        }
-
-        // Create tne temp folder if it doesn't exist
-        // This will be used to write output files from ProcessBuilder commands
-        // called elsewhere in this class
-        val tempDir = "${dbPath}/log"
-        Files.createDirectories(Paths.get(tempDir))
-
-        if (File(dataset).exists()  && Files.isDirectory(Paths.get(dataset))){
-            // check if is a tiledb dataset
-            var builder = ProcessBuilder("conda","run","-n","phgv2-conda","tiledbvcf","stat","--uri",dataset)
-            var redirectOutput = tempDir + "/tiledb_statURI_output.log"
-            var redirectError = tempDir + "/tiledb_statURI_error.log"
-            builder.redirectOutput( File(redirectOutput))
-            builder.redirectError( File(redirectError))
-
-            // verify if the output.log contains "Version"
-            // if not, then the URI is not a tiledbvcf URI
-            myLogger.info("begin Command:" + builder.command().joinToString(" "))
-
-            try {
-                var process = builder.start()
-                var error = process.waitFor()
-                if (error != 0) {
-                    myLogger.error("LoadTiledbH tiledb stat returned error code $error")
-                    throw IllegalArgumentException("Error: URI is not a tiledb URI folder created via the tiledb create command: ${error}")
-                }
-
-            } catch (exc: Exception) {
-                myLogger.error("Error: could not run tiledb stat on ${uri}.")
-                throw IllegalArgumentException("Error running ProcessBuilder to stat tiledb URI: ${exc}")
-            }
-
-            myLogger.info("Using existing TileDB datasets previously created in folder $dbPath.")
-            return true
-        } else {
-            myLogger.info("TileDB datasets not found in folder $dbPath. Please run InitDB to create the datasets.")
-            return false
-        }
     }
 
     fun loadVCFToTiledb(vcfList:List<String>, uri:String, threads:String) {
